requirements.txt
google-cloud-pubsub
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
playsound